+++ pkg/util/validation/errors.go
@@ -26,6 +26,17 @@ func (e *genericValidationError) Error() string {
 	return fmt.Sprintf(e.message, e.cause, formatLabelSet(e.series))
 }
 
+// labelValueTooLongError is a customized ValidationError, in that the cause and the series are
+// are formatted in different order in Error.
+type labelValueTooLongError struct {
+	labelValue string
+	series     []cortexpb.LabelAdapter
+}
+
+func (e *labelValueTooLongError) Error() string {
+	return fmt.Sprintf("label value too long for metric: %.200q label value: %.200q", formatLabelSet(e.series), e.labelValue)
+}
+
 func newLabelNameTooLongError(series []cortexpb.LabelAdapter, labelName string) ValidationError {
 	return &genericValidationError{
 		message: "label name too long: %.200q metric %.200q",
@@ -35,10 +46,9 @@ func newLabelNameTooLongError(series []cortexpb.LabelAdapter, labelName string) 
 }
 
 func newLabelValueTooLongError(series []cortexpb.LabelAdapter, labelValue string) ValidationError {
-	return &genericValidationError{
-		message: "label value too long: %.200q metric %.200q",
-		cause:   labelValue,
-		series:  series,
+	return &labelValueTooLongError{
+		labelValue: labelValue,
+		series:     series,
 	}
 }